bot/requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# HTTP Client
requests==2.31.0

# Environment
python-decouple==3.8

# Async Support
aiohttp==3.9.1
asyncio==3.4.3

# Utilities
python-dateutil==2.8.2
Pillow==10.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

